Create TransformationWrapperController
Moves transformation wrapper entry point logic to the
TransformationWrapper controller and removes the entry point file.
