+++ services_controllers.yml
@@ -408,6 +408,15 @@ services:
       template: '@template'
       transformations: '@transformations'
 
+  PhpMyAdmin\Controllers\TransformationWrapperController:
+    class: 'PhpMyAdmin\Controllers\TransformationWrapperController'
+    arguments:
+      response: '@response'
+      dbi: '@dbi'
+      template: '@template'
+      transformations: '@transformations'
+      relation: '@relation'
+
   PhpMyAdmin\Controllers\UserPasswordController:
     class: 'PhpMyAdmin\Controllers\UserPasswordController'
     arguments:
